requirements.txt
click==8.0.3
colorama==0.4.4
Flask==2.0.2
gunicorn==20.1.0
imbalanced-learn==0.8.1
imblearn==0.0
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
numpy==1.21.5
pandas==1.3.5
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
threadpoolctl==3.0.0
Werkzeug==2.0.2
